fix: add required migration (#856)
I added the required migration for the new deployments, and made the
`BumpStorageVersion` type a bit more generic to work with any pallets
that might need the same treatment in the future.

I tried the new migrations with both Peregrine (new runtime) and
Spiritnet (previous runtime), and I got no errors.