feat: add new DID runtime APIs to aid with DID deletion (#852)
Fixes https://github.com/KILTprotocol/ticket/issues/3742.

The PR introduces two new runtime APIs: one that returns an `enum` of
"linked resources", which some clients might find useful to understand
why a DID deletion failed, and one that returns a list of
`RuntimeCall`s, that could be extended with a `did.delete()` call, put
together into a `utility.batch*`, and submitted to the chain, and be
sure the DID deletion will work as expected (as long as all the DID
origin requirements are met).