fix: remove dotnames (#862)
This pull request removes the `unique-linking` and `dot-names` pallet
implementations. However, the pallet changes are retained, as they might
be useful in the future.

All related implementations, including the runtime API, runtime commons,
benchmarking, migrations, and miscellaneous components, are removed.

For Peregrine, a migration is introduced to remove the already deployed
pallets. For Spiritnet, the pallet implementation in the runtime
construction is removed without additional comments.