feat: Permissioned collators (#882)
## fixes [#3822](https://github.com/KILTprotocol/ticket/issues/3822)

This PR introduces permissioned collators to the Peregrine and Spirtinet
runtimes. The council or governance is now allowed to add and remove
members through the newly deployed pallet-membership. This pallet is
then used by a new session manager implementation. The old session
manager implementation remains untouched—it's only moved out of the
runtime.

After the migration, the parachain-staking pallet becomes essentially
obsolete, as collators no longer receive rewards for block production.

A follow-up PR will reduce rewards in the parachain-staking pallet
accordingly.

The round information (which determines session length) still resides in
the parachain-staking pallet. Before it can be removed, a new source for
round length must be implemented—either directly in the runtime or via a
custom pallet that includes collator information.

### How to test
- Navigate to the zombienet directory and spin up a Peregrine runtime.
- Perform the runtime upgrade (sudo -> system -> set_code).
- Add a new collator via sudo or start a council motion (sudo ->
permissionedCollators -> add_member).
- Force a new round (sudo -> parachainStaking -> forceNewRound).
- Verify in the session pallet that the new collator is queued. (To use
a new collator, spin up a node with the correct keys. Using an existing
collator is usually easier.)
- Force another new round.
- Only the new collator should be producing blocks now.